Add six and packaging to the hidden imports in the spec file.
As later versions of `setuptools` appear to be using `six`, we need to
make sure it is included in generated binary. However, it is not
imported by anything in statuspage. So, we need to convince PyInstaller
to include this anyway. Adding it to the hidden imports does this.

Also, `packaging` appears to need importing as well. So, here we note
that it is a hidden import too. That way PyInstaller will make sure to
pick it up as well.