Fix for issues not being marked as resolved
The "resolved" label is not in list of labels created by the system, so that caused that 
```python
if not affected_systems or severity is None:
```
to have `severity = None`, which basically made it ignore closed issues with no severity status (since you remove the label), after resolving an issue).

Now it works as intended. User removes severity label, closes the issue - statuspage generates proper resolved issue in the list.